More on refactoring towards DI

+++ SurveyWeb/Repository/TaskRepository.cs
@@ -88,9 +88,12 @@ namespace SurveyWeb.Repository
             {
                 if (selectedUsers.Count(userId => userId == item.Id) == 0)
                 {
-                    var toDelete = _context.EventUsers.First(eventUser => eventUser.EventId == taskId && eventUser.UserId == item.Id);
+                    var toDelete = _context.EventUsers.FirstOrDefault(eventUser => eventUser.EventId == taskId && eventUser.UserId == item.Id);
 
-                    _context.EventUsers.Remove(toDelete);
+                    if (toDelete != default(EventUser))
+                    {
+                        _context.EventUsers.Remove(toDelete);
+                    }
                 }
             }
